Transferred test.js to server.js
